+++ TableParts/TableParts/AppDelegate.cs
@@ -1,30 +1,21 @@
-using System;
+﻿using Foundation;
 using UIKit;
-using Foundation;
 
-namespace Example_TableParts
+namespace TableParts
 {
-	[Register("AppDelegate")]
-	public class AppDelegate : UIApplicationDelegate
-	{
-		protected UIWindow window;
-		protected Example_TableParts.Screens.iPhone.Home.HomeScreen iPhoneHome;
+    // The UIApplicationDelegate for the application. This class is responsible for launching the
+    // User Interface of the application, as well as listening (and optionally responding) to application events from iOS.
+    [Register("AppDelegate")]
+    public class AppDelegate : UIApplicationDelegate
+    {
+        public override UIWindow Window { get; set; }
 
-		public override bool FinishedLaunching (UIApplication app, NSDictionary options)
-		{
-			// create our window
-			window = new UIWindow (UIScreen.MainScreen.Bounds);
-			window.MakeKeyAndVisible ();
+        public override bool FinishedLaunching(UIApplication application, NSDictionary launchOptions)
+        {
+            // Override point for customization after application launch.
+            // If not required for your application you can safely delete this method
 
-			// create the home screen and modify it's frame
-			iPhoneHome = new Example_TableParts.Screens.iPhone.Home.HomeScreen();
-			iPhoneHome.View.Frame = new CoreGraphics.CGRect(0, UIApplication.SharedApplication.StatusBarFrame.Height, UIScreen.MainScreen.ApplicationFrame.Width, UIScreen.MainScreen.ApplicationFrame.Height);
-
-			window.RootViewController = iPhoneHome;
-
-			//
-			return true;
-		}
-
-	}
+            return true;
+        }
+    }
 }
