+++ TableParts/TableParts/Info.plist
@@ -1,31 +1,40 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>CFBundleDisplayName</key>
+	<key>CFBundleName</key>
 	<string>TableParts</string>
-	<key>CFBundleIconFiles</key>
-	<array>
-		<string>Resources/icon-57.png</string>
-		<string>Resources/icon-114.png</string>
-		<string>Resources/icon-72.png</string>
-		<string>Resources/icon-144.png</string>
-		<string>Resources/icon-29.png</string>
-		<string>Resources/icon-58.png</string>
-		<string>Resources/icon-50.png</string>
-		<string>Resources/icon-100.png</string>
-	</array>
+	<key>CFBundleIdentifier</key>
+	<string>com.xamarin.TableParts</string>
+	<key>CFBundleShortVersionString</key>
+	<string>1.0</string>
+	<key>CFBundleVersion</key>
+	<string>1.0</string>
+	<key>LSRequiresIPhoneOS</key>
+	<true/>
 	<key>MinimumOSVersion</key>
-	<string>9.0</string>
+	<string>12.1</string>
 	<key>UIDeviceFamily</key>
 	<array>
 		<integer>1</integer>
+		<integer>2</integer>
 	</array>
-	<key>CFBundleIdentifier</key>
-	<string>com.xamarin.TableParts</string>
+	<key>UILaunchStoryboardName</key>
+	<string>LaunchScreen</string>
+	<key>UIMainStoryboardFile</key>
+	<string>Main</string>
+	<key>UIRequiredDeviceCapabilities</key>
+	<array>
+		<string>armv7</string>
+	</array>
 	<key>UISupportedInterfaceOrientations</key>
 	<array>
 		<string>UIInterfaceOrientationPortrait</string>
+		<string>UIInterfaceOrientationPortraitUpsideDown</string>
+		<string>UIInterfaceOrientationLandscapeLeft</string>
+		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
+	<key>XSAppIconAssets</key>
+	<string>Assets.xcassets/AppIcon.appiconset</string>
 </dict>
 </plist>
