+++ TableParts/TableParts/Main.cs
@@ -1,0 +1,15 @@
+﻿using UIKit;
+
+namespace TableParts
+{
+    public class Application
+    {
+        // This is the main entry point of the application.
+        static void Main(string[] args)
+        {
+            // if you want to use a different Application Delegate class from "AppDelegate"
+            // you can specify it here.
+            UIApplication.Main(args, null, "AppDelegate");
+        }
+    }
+}
