+++ TableParts/TableParts2/Resources/Default-Landscape@2x~ipad.png
